refactor: relocate test-only helpers out of src/commands (tier A6-A8, #338)
Continues the relocation begun in the prior commit. After this commit,
the PENDING_MIGRATION allow-list in tests/unit/test-import-boundary.test.ts
contains a single entry: tests/unit/identity.test.ts (tier B — deferred,
see PR description and #338 for migration plan).

Tier A6: src/commands/completion.ts -> src/completion.ts
  - All shell-completion rendering, install, detect, refresh helpers
    move to top-level src/completion.ts. The thin
    src/commands/completion.ts wrapper now contains only the
    completionCommand defineCommand handler, which imports its
    implementation from ../completion.ts.
  - src/index.ts updated to import refreshCompletionsIfStale from the
    new location.

Tier A7: src/commands/help.ts -> src/help.ts (whole-file move)
  - help.ts has no defineCommand wrapper (help is invoked directly from
    src/index.ts), so the whole file moves. __dirname-based package.json
    lookup updated from ../../package.json to ../package.json.
  - src/commands/mcp-proxy.ts and src/index.ts updated to import
    getVersion / showHelp / etc. from ../help.ts and ./help.ts.

Tier A8: src/commands/deployments.ts -> src/deployments.ts (whole-file move)
  - The deployCommand defineCommand wrapper and the shared
    deployResources helper both move; the helper file's only
    cross-package consumer is src/commands/watch.ts (now imports from
    ../deployments.ts) and src/command-dispatch.ts (now imports
    deployCommand from ./deployments.ts).
  - tests/unit/deploy-error-paths.test.ts DEPLOYMENTS_TS path constant
    updated. Doc comments referencing the old path in
    tests/unit/normalize-to-error.test.ts, round-3-baseline.test.ts,
    and tests/integration/process-instances.test.ts updated for
    consistency.

Tier B (unit/identity.test.ts) deferred. The identity tests build a
CommandContext directly via buildCtx(), invoke defineCommand outputs as
functions, and assert against captured stdout/stderr. Migrating these
56 tests to a c8() subprocess pattern is a real rewrite, not a
mechanical relocation, and is tracked in #338 for a follow-up PR.

Refs #338
