+++ src/help.ts
@@ -12,14 +12,14 @@ import {
 	GLOBAL_FLAGS,
 	RESOURCE_ALIASES,
 	SEARCH_FLAGS,
-} from "../command-registry.ts";
-import { getLogger } from "../logger.ts";
+} from "./command-registry.ts";
+import { getLogger } from "./logger.ts";
 import {
 	executePluginCommand,
 	getPluginCommandsInfo,
 	isPluginCommand,
 	type PluginCommandInfo,
-} from "../plugin-loader.ts";
+} from "./plugin-loader.ts";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -197,7 +197,7 @@ function buildHelpJson(
  * Get package version
  */
 export function getVersion(): string {
-	const packagePath = join(__dirname, "../../package.json");
+	const packagePath = join(__dirname, "../package.json");
 	const packageJson = JSON.parse(readFileSync(packagePath, "utf-8"));
 	return packageJson.version;
 }
