+++ tests/integration/deploy.test.ts
@@ -7,8 +7,8 @@ import assert from "node:assert";
 import { existsSync, unlinkSync } from "node:fs";
 import { join } from "node:path";
 import { beforeEach, describe, test } from "node:test";
-import { deployResources as deploy } from "../../src/commands/deployments.ts";
 import { getUserDataDir } from "../../src/config.ts";
+import { deployResources as deploy } from "../../src/deployments.ts";
 
 describe("Deployment Integration Tests (requires Camunda 8 at localhost:8080)", () => {
 	beforeEach(() => {
